+++ bigml/tests/compare_predictions_steps.py
@@ -42,6 +42,14 @@ def i_retrieve_a_list_of_remote_models(step, tag):
         world.api.list_models(query_string="tags__in=%s" % tag)['objects']]
 
 
+#@step(r'I retrieve a list of remote logistic regression tagged with "(.*)"')
+def i_retrieve_a_list_of_remote_logistic_regressions(step, tag):
+    world.list_of_models = [ \
+        world.api.get_logistic_regression(model['resource']) for model in
+        world.api.list_logistic_regressions( \
+            query_string="tags__in=%s" % tag)['objects']]
+
+
 #@step(r'I create a local model from a "(.*)" file$')
 def i_create_a_local_model_from_file(step, model_file):
     world.local_model = Model(res_filename(model_file))